combine everything to LayerNode
