+++ packages/core/test/effect/layer-node/node-build.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "bun:test"
 import { Context, Effect, Layer, LayerMap, Option } from "effect"
-import { LayerNode, LayerNodeTree } from "@opencode-ai/core/effect/layer-node"
+import { LayerNode } from "@opencode-ai/core/effect/layer-node"
 import { Node } from "@opencode-ai/core/effect/app-node"
 import { NodeBuild } from "@opencode-ai/core/effect/app-node-builder"
 import { Location } from "@opencode-ai/core/location"
@@ -62,7 +62,7 @@ describe("node build", () => {
     }) as unknown as Effect.Effect<LayerMap.LayerMap<Location.Ref, LocationServices, LocationError>, never, CycleB>
     const mapLayer = Layer.effect(LocationServiceMap.Service, mapEffect)
     const map = Node.makeGlobalNode({ service: LocationServiceMap.Service, layer: mapLayer, deps: [b] })
-    const graph = LayerNodeTree.bind(LayerNode.group([a]), LocationServiceMap.node, map)
+    const graph = LayerNode.bind(LayerNode.group([a]), LocationServiceMap.node, map)
 
     expect(() => NodeBuild.build(graph)).toThrow("Cycle detected in layer tree")
   })
