+++ packages/opencode/test/git/git.test.ts
@@ -1,5 +1,5 @@
 import { $ } from "bun"
-import { LayerNodeTree } from "@opencode-ai/core/effect/layer-node"
+import { LayerNode } from "@opencode-ai/core/effect/layer-node"
 import { describe, expect } from "bun:test"
 import fs from "fs/promises"
 import path from "path"
@@ -10,7 +10,7 @@ import { tmpdir } from "../fixture/fixture"
 import { testEffect } from "../lib/effect"
 
 const weird = process.platform === "win32" ? "space file.txt" : "tab\tfile.txt"
-const it = testEffect(LayerNodeTree.compile(LayerNode.group([Git.node])))
+const it = testEffect(LayerNode.compile(LayerNode.group([Git.node])))
 
 const scopedTmpdir = (options?: Parameters<typeof tmpdir>[0]) =>
   Effect.acquireRelease(
