+++ packages/server/src/routes.ts
@@ -1,5 +1,5 @@
 import { Database } from "@opencode-ai/core/database/database"
-import { LayerNode, LayerNodeTree } from "@opencode-ai/core/effect/layer-node"
+import { LayerNode } from "@opencode-ai/core/effect/layer-node"
 import { httpClient } from "@opencode-ai/core/effect/app-node-platform"
 import { NodeBuild } from "@opencode-ai/core/effect/app-node-builder"
 import { EventV2 } from "@opencode-ai/core/event"
@@ -50,7 +50,7 @@ export function createEmbeddedRoutes() {
 
 function makeRoutes<AuthError, AuthServices>(auth: Layer.Layer<ServerAuth.Config, AuthError, AuthServices>) {
   const serviceLayer = NodeBuild.build(
-    LayerNodeTree.bind(applicationServices, SessionExecution.node, SessionExecutionLocal.node),
+    LayerNode.bind(applicationServices, SessionExecution.node, SessionExecutionLocal.node),
   )
 
   return HttpApiBuilder.layer(Api, { openapiPath: "/openapi.json" }).pipe(
